Module oak_containers_orchestrator_client::proto::oak::containers

Modules

launcher_client

Generated client implementations.

orchestrator_client

Generated client implementations.

v1

Structs

GetApplicationConfigResponse

GetCryptoContextRequest

GetCryptoContextResponse

GetImageResponse

As images can be large (hundreds of megabytes), the launcher chunks up the response into smaller pieces to respect proto/gRPC limits. The image needs to be reassembled in the stage1 or the orchestrator.

SendAttestationEvidenceRequest